edk2-email-archive-webhook: Add repo scoped statistics
* Move repo scoped statistics into WebhookConfigtation
* Compute service scoped stats by summing repo scoped stats
* Remove option to reset statistics at entire service scope
* Do not collect stats on CUSTOM events internal to the service
* Add option to reset repo scoped statistics
* Simplify HTML by using <button> instead of <input>.  <button>
  more easily support multi-line button labels.

+++ GithubRequest/GithubRequest.py
@@ -377,6 +377,9 @@ class GithubRequest(object):
 
     def DispatchGithubRequest(self):
         try:
+            if self.event=='CUSTOM' and self.action=='ResetStatistics':
+                self.webhookconfiguration.ResetStatistics()
+                return 200, 'successfully processed %s event with action %s' % (self.event, self.action)
             if self.event=='CUSTOM' and self.action=='DeleteRepositoryCache':
                 return self.DeleteRepositoryCache()
             if self.event=='CUSTOM' and self.action=='ClearLogs':
@@ -417,7 +420,8 @@ class GithubRequest(object):
             self.payload,
             datetime.now()
             ))
-        WebhookStatistics.query.all()[0].RequestQueued()
+        if self.event != 'CUSTOM':
+            self.webhookconfiguration.RequestQueued()
         return 202, 'Event %s with action %s queued for processing' % (self.event, self.action)
 
     def ProcessGithubRequest(self):
